+++ otynium.rb
@@ -1,260 +1,238 @@
-require 'json'
+require './colorscheme.rb'
 
-class Vim
-  def initialize(name, palette)
-    @text = ''
-    @text << "hi clear\n"
-    @text << "if version > 580\n"
-    @text << "  hi clear \n"
-    @text << "  if exists(\"syntax_on\") \n"
-    @text << "    syntax reset \n"
-    @text << "  endif \n"
-    @text << "endif\n"
-    @text << "\n"
-    @text << "set background=dark\n"
-    @text << "set termguicolors\n"
-    @text << "let g:colors_name = '#{name}'\n"
-    @palette = palette
-  end
+color_otynium = {
+  red:     "\#ef7070",
+  red2:    "\#8a1a1a",
+  orange:  "\#e8a368",
+  yellow:  "\#d8cf6c",
+  yellow2: "\#e2d86a",
+  green:   "\#85c178",
+  green2:  "\#5cc46b",
+  cyan:    "\#64bac9",
+  blue:    "\#6882e2",
+  purple:  "\#b281d1",
+  gray1:   "\#222c33",
+  gray2:   "\#314047",
+  gray3:   "\#506872",
+  gray4:   "\#5f7b87",
+  gray5:   "\#7395a3",
+  white:   "\#a9afb2",
+  black:   "\#050a26",
+}
 
-  def link(from, to)
-    @text << "hi link #{from} #{to}\n"
-  end
+color_otynium256 = {
+  red:     "\#d75f5f",
+  red2:    "\#870000",
+  orange:  "\#ffaf5f",
+  yellow:  "\#d7d75f",
+  yellow2: "\#d7d700",
+  green:   "\#87af87",
+  green2:  "\#5faf5f",
+  cyan:    "\#5fafd7",
+  blue:    "\#5f87ff",
+  purple:  "\#af87ff",
+  gray1:   "\#262626",
+  gray2:   "\#303030",
+  gray3:   "\#4e4e4e",
+  gray4:   "\#585858",
+  gray5:   "\#767676",
+  white:   "\#a8a8a8",
+  black:   "\#00005f",
+}
 
-  def hi(group, color_name)
-    @text << "hi #{group}"
-    @text << " guifg=#{@palette.fg(color_name)}" if @palette.fg?(color_name)
-    @text << " guibg=#{@palette.bg(color_name)}" if @palette.bg?(color_name)
-
-    unless @palette.attrs(color_name).empty?
-      attr = @palette.attrs(color_name).map {|attr| attr.to_s}.join(',')
-      @text << " gui=#{attr} cterm=#{attr}" 
-    end
-    @text << "\n"
-  end
-
-  def output
-    @text
-  end
-end
-
-class Palette
-  def initialize(color)
-    @color = color
-    @colors = {}
-  end
-
-  def choose_color(color)
-    return @color[color] if color.is_a? Symbol
-    color
-  end
-
-  def hi(name, fg, bg, attrs)
-    @colors[name] = { fg: choose_color(fg), bg: choose_color(bg), attr: attrs }
-  end
-
-  def fg(name)
-    @colors[name][:fg]
-  end
-
-  def fg?(name)
-    !@colors[name][:fg].nil?
-  end
-
-  def bg(name)
-    @colors[name][:bg]
-  end
-
-  def bg?(name)
-    !@colors[name][:bg].nil?
-  end
-
-  def italic?(name)
-    @colors[name][:attr].include? :italic
-  end
-
-  def underline?(name)
-    @colors[name][:attr].include? :underline
-  end
-
-  def bold?(name)
-    @colors[name][:attr].include? :bold
-  end
-
-  def none?(name)
-    @colors[name][:attr].include? :none
-  end
-
-  def attrs(name)
-    @colors[name][:attr]
-  end
-end
-
-class TokenColor
-  def initialize(name, palette)
-    @name = name
-    @palette = palette
-    @settings = {}
-  end
-
-  def scope(scope)
-    @scope = scope
-  end
-
-  def style(style)
-    @settings[:fontStyle] = style
-  end
-
-  def fore(fg)
-    @settings[:foreground] = @palette.fg(fg)
-  end
-
-  def back(bg)
-    @settings[:background] = color(@palette, bg)
-  end
-
-  def json
-    {
-      name: @name,
-      scope: @scope,
-      settings: @settings
-    }
-  end
-end
-
-class VsCode
-  def initialize(name, palette, type)
-    @json = {
-      name: name,
-      type: type,
-      colors: {},
-      tokenColors: [],
-    }
-    @palette = palette
-  end
-
-  def background(bg)
-    @json[:colors]['editor.background'] = @palette.bg(bg)
-  end
-
-
-  def foreground(fg)
-    @json[:colors]['editor.foreground'] = @palette.fg(fg)
-  end
-
-  def barbadge_background(bg)
-    @json[:colors]['activityBarBadge.background'] = @palette.bg(bg)
-  end
-
-  def sidebartitle_foregraund(fg)
-    @json[:colors]['sideBarTitle.foreground'] = @palette.fg(fg)
-  end
-
-  def token(name, style, scope)
-    settings = {}
-    settings[:foreground] = @palette.fg(style) if @palette.fg(style)
-    settings[:fontStyle] = 'underline' if @palette.attrs(style).include? :underline
-    settings[:fontStyle] = 'italic' if @palette.attrs(style).include? :italic
-    settings[:fontStyle] = 'bold' if @palette.attrs(style).include? :bold
-    @json[:tokenColors] << {
-      name: name,
-      scope: scope,
-      settings: settings
-    }
-  end
-
-  def output
-    JSON.pretty_generate @json
+def gen_style(color)
+  styles(color) do
+    hi :Normal              , :white   , :gray1  , []
+    hi :NonText             , :gray2   , nil     , []
+    hi :Cursor              , :gray2   , :gray3  , [:none]
+    hi :CursorLine          , nil      , :gray2  , [:none]
+    hi :CursorLineNr        , :cyan    , :gray2  , [:none]
+    hi :CursorColumn        , nil      , :gray2  , []
+    hi :LineNr              , :gray2   , nil     , []
+    hi :StatusLine          , :gray5   , :gray2  , []
+    hi :StatusLineNC        , :gray1   , :gray4  , []
+    hi :StatusLineTerm      , :gray5   , :gray2  , []
+    hi :StatusLineTermNC    , :gray1   , :gray4  , []
+    hi :Title               , :green   , nil     , []
+    hi :Directory           , :blue    , nil     , []
+    hi :Pmenu               , :white   , :black  , []
+    hi :PmenuSel            , :white   , :gray2  , []
+    hi :PmenuSbar           , :gray3   , nil     , []
+    hi :PmenuThumb          , :gray4   , nil     , []
+    hi :Search              , :gray2   , :yellow , []
+    hi :IncSearch           , :gray1   , :yellow , []
+    hi :DiffAdd             , :cyan    , :gray2  , []
+    hi :DiffChange          , :yellow2 , :gray2  , []
+    hi :DiffDelete          , :red     , :gray2  , []
+    hi :DiffText            , :green2  , :gray2  , []
+    hi :SpellBad            , :red     , nil     , []
+    hi :SpellCap            , :blue    , nil     , []
+    hi :SpellLocal          , :green   , nil     , []
+    hi :SpellRare           , :gray2   , :yellow , []
+    hi :SpecialKey          , :gray2   , nil     , []
+    hi :WildMenu            , :blue    , nil     , []
+    hi :Folded              , :gray4   , :gray1  , []
+    hi :FoldColumn          , :gray4   , :gray1  , []
+    hi :VertSplit           , :gray4   , nil     , []
+    hi :TabLine             , :gray4   , :gray2  , []
+    hi :TabLineFill         , :gray4   , :gray2  , []
+    hi :TabLineSel          , :orange  , :gray3  , []
+    hi :ColorColumn         , nil      , :gray1  , []
+    hi :SignColumn          , :gray5   , :gray1  , []
+    hi :Question            , :cyan    , nil     , []
+    hi :Visual              , :gray5   , :gray3  , []
+    hi :VisualNOS           , :gray5   , :gray3  , []
+    hi :ModeMsg             , :yellow  , nil     , []
+    hi :MoreMsg             , :yellow  , nil     , []
+    hi :WarningMsg          , :red     , nil     , []
+    hi :ErrorMsg            , nil      , :red2   , []
+    hi :Constant            , :orange  , nil     , []
+    hi :String              , :green   , nil     , []
+    hi :Character           , :green   , nil     , []
+    hi :Number              , :orange  , nil     , []
+    hi :Float               , :orange  , nil     , []
+    hi :Boolean             , :orange  , nil     , []
+    hi :Identifier          , :red     , nil     , [:none]
+    hi :Function            , :cyan    , nil     , []
+    hi :Statement           , :purple  , nil     , []
+    hi :Conditional         , :purple  , nil     , []
+    hi :Repeat              , :purple  , nil     , []
+    hi :Label               , :purple  , nil     , []
+    hi :Operator            , :blue    , nil     , []
+    hi :Keyword             , :purple  , nil     , []
+    hi :Exception           , :yellow  , nil     , []
+    hi :Type                , :yellow  , nil     , []
+    hi :StorageClass        , :orange  , nil     , []
+    hi :Structure           , :orange  , nil     , []
+    hi :Typedef             , :yellow  , nil     , []
+    hi :PreProc             , :cyan    , nil     , []
+    hi :Include             , :cyan    , nil     , []
+    hi :Define              , :cyan    , nil     , []
+    hi :Macro               , :cyan    , nil     , []
+    hi :PreCondit           , :blue    , nil     , []
+    hi :Special             , :cyan    , nil     , []
+    hi :SpecialChar         , :cyan    , nil     , []
+    hi :Tag                 , :blue    , nil     , []
+    hi :Delimiter           , :blue    , nil     , []
+    hi :SpecialComment      , :cyan    , nil     , []
+    hi :Debug               , :red     , nil     , []
+    hi :Underlined          , :red     , nil     , []
+    hi :Comment             , :gray4   , nil     , []
+    hi :Todo                , :gray1   , :yellow , []
+    hi :Ignore              , :gray3   , nil     , []
+    hi :Error               , nil      , :red2   , []
+    hi :MatchParen          , :gray2   , :cyan   , []
+    hi :Warning             , :orange  , nil     , []
+    hi :hint                , nil      , :gray3  , []
+    hi :hintMsg             , :gray3   , nil     , []
+    hi :Info                , nil      , :cyan   , []
+    hi :InfoMsg             , :cyan    , nil     , []
+    hi :CocErrorhighlight   , nil      , nil     , [:underline]
+    hi :CocWarninghighlight , nil      , nil     , [:underline]
+    hi :Cochinthighlight    , nil      , nil     , [:underline]
+    hi :CocInfohighlight    , nil      , nil     , [:underline]
+    hi :TSVariable          , :cyan    , nil     , []
+    hi :TSStrong            , :white   , :gray1  , [:bold]
+    hi :TSStrong            , :white   , :gray1  , [:underline]
+    hi :TSLiteral           , :orange  , nil     , []
+    hi :TSURI               , :orange  , nil     , []
+    hi :TSTagDelimiter      , :gray4   , nil     , []
+    hi :markdownH1          , :white   , nil     , [:bold]
+    hi :markdownH2          , :white   , nil     , [:bold]
+    hi :markdownH3          , :white   , nil     , [:bold]
+    hi :markdownH4          , :white   , nil     , [:bold]
+    hi :markdownH5          , :white   , nil     , [:bold]
+    hi :markdownH6          , :white   , nil     , [:bold]
+    hi :markdownBold        , :white   , nil     , [:bold]
+    hi :markdownItalic      , :white   , nil     , [:italic]
+    hi :markdownBoldItalic  , :white   , nil     , [:italic     , :bold]
   end
 end
 
-def vscode_colorscheme(name, palette, type)
-  vscode = VsCode.new(name, palette, type)
-  vscode.instance_eval do
-    background :Normal
-    foreground :Normal
-    barbadge_background :Normal
-    sidebartitle_foregraund :Normal
-    token 'Comment', :Comment, ['comment', 'punctuation.definition.comment']
-    token 'Variables', :TSVariable, ['variable', 'string constant.other.placeholder']
-    token 'Colors', :Normal, ['constant.other.color']
-    token 'Invalid', :Error, ['invalid', 'invalid.illegal']
-    token 'Keyword', :Keyword, ['keyword', 'storage.type', 'storage.modifier']
-    token 'Storage', :StorageClass, ['storage.type', 'storage.modifier']
-    # TODO
-    token 'Operator, Misc', :Operator, [
-      "keyword.control",
-      "constant.other.color",
-      "punctuation",
-      "meta.tag",
-      "punctuation.definition.tag",
-      "punctuation.separator.inheritance.php",
-      "punctuation.definition.tag.html",
-      "punctuation.definition.tag.begin.html",
-      "punctuation.definition.tag.end.html",
-      "punctuation.section.embedded",
-      "keyword.other.template",
-      "keyword.other.substitution"
-		]
-    token 'Block Level Variables, Other Variables, String Link, Tag', :Tag, [
-      "entity.name.tag",
-      "meta.tag.sgml",
-      "markup.deleted.git_gutter"
-    ]
-    token 'Function, Special Method', :Function, [
-      'entity.name.function', 
-      "meta.function-call",
-      "variable.function",
-      "support.function",
-      "keyword.other.special-method"
-    ]
-    token 'Number', :Number, ['constant.numeric']
-    token 'Constant', :Constant, ['constant.language', 'support.constant']
-    token 'Function Argument, Tag Attribute, Embedded', :Constant, [
-      "constant.character",
-      "constant.escape",
-      "variable.parameter",
-      "keyword.other.unit",
-      "keyword.other"
-    ]
-    token 'String, Symbols, Inherited Class, Markup Heading', :String, [
-      "string",
-      "constant.other.symbol",
-      "constant.other.key",
-      "entity.other.inherited-class",
-      "markup.heading",
-      "markup.inserted.git_gutter",
-      "meta.group.braces.curly constant.other.object.key.js string.unquoted.label.js"
-    ]
-    token 'Class, Support', :Keyword, [
-      "entity.name",
-      "support.type",
-      "support.class",
-      "support.other.namespace.use.php",
-      "meta.use.php",
-      "support.other.namespace.php",
-      "markup.changed.git_gutter",
-      "support.type.sys-types"
-		]
-    token 'Entity Types', :Typedef, [
-      "support.type"
-    ]
-    token 'CSS Class and Support', :Typedef, [
-      "source.css support.type.property-name",
-      "source.sass support.type.property-name",
-      "source.scss support.type.property-name",
-      "source.less support.type.property-name",
-      "source.stylus support.type.property-name",
-      "source.postcss support.type.property-name",
-    ]
-  end
-  vscode.output
+style_otynium = gen_style(color_otynium)
+style_otynium256 = gen_style(color_otynium256)
+
+vscode = vscode_colorscheme('otynium', style_otynium, 'dark') do
+  background :Normal
+  foreground :Normal
+  barbadge_background :Normal
+  sidebartitle_foregraund :Normal
+  token ['Comment'], :Comment, ['comment', 'punctuation.definition.comment']
+  token ['Variables'], :TSVariable, ['variable', 'string constant.other.placeholder']
+  token ['Colors'], :Normal, ['constant.other.color']
+  token ['Invalid'], :Error, ['invalid', 'invalid.illegal']
+  token ['Keyword'], :Keyword, ['keyword', 'storage.type', 'storage.modifier']
+  token ['Storage'], :StorageClass, ['storage.type', 'storage.modifier']
+  token ['Operator', 'Misc'], :Operator, [
+    "keyword.control",
+    "constant.other.color",
+    "punctuation",
+    "meta.tag",
+    "punctuation.definition.tag",
+    "punctuation.separator.inheritance.php",
+    "punctuation.definition.tag.html",
+    "punctuation.definition.tag.begin.html",
+    "punctuation.definition.tag.end.html",
+    "punctuation.section.embedded",
+    "keyword.other.template",
+    "keyword.other.substitution"
+  ]
+  token ['Block Level Variables', 'Other Variables', 'String Link, Tag'], :Tag, [
+    "entity.name.tag",
+    "meta.tag.sgml",
+    "markup.deleted.git_gutter"
+  ]
+  token ['Function', 'Special Method'], :Function, [
+    'entity.name.function', 
+    "meta.function-call",
+    "variable.function",
+    "support.function",
+    "keyword.other.special-method"
+  ]
+  token ['Number'], :Number, ['constant.numeric']
+  token ['Constant'], :Constant, ['constant.language', 'support.constant']
+  token ['Function Argument', 'Tag Attribute, Embedded'], :Constant, [
+    "constant.character",
+    "constant.escape",
+    "variable.parameter",
+    "keyword.other.unit",
+    "keyword.other"
+  ]
+  token ['String, Symbols', 'Inherited Class', 'Markup Heading'], :String, [
+    "string",
+    "constant.other.symbol",
+    "constant.other.key",
+    "entity.other.inherited-class",
+    "markup.heading",
+    "markup.inserted.git_gutter",
+    "meta.group.braces.curly constant.other.object.key.js string.unquoted.label.js"
+  ]
+  token ['Class, Support'], :Keyword, [
+    "entity.name",
+    "support.type",
+    "support.class",
+    "support.other.namespace.use.php",
+    "meta.use.php",
+    "support.other.namespace.php",
+    "markup.changed.git_gutter",
+    "support.type.sys-types"
+      ]
+  token ['Entity Types'], :Typedef, [
+    "support.type"
+  ]
+  token ['CSS Class and Support'], :Typedef, [
+    "source.css support.type.property-name",
+    "source.sass support.type.property-name",
+    "source.scss support.type.property-name",
+    "source.less support.type.property-name",
+    "source.stylus support.type.property-name",
+    "source.postcss support.type.property-name",
+  ]
 end
 
 
-def vim_colorscheme(name, palette)
-  vim = Vim.new(name, palette)
-  vim.instance_eval do
+def gen_vim(name, style)
+  vim_colorscheme(name, style) do
     hi 'Normal'           , :Normal
     hi 'NonText'          , :NonText
     hi 'Cursor'           , :Cursor
@@ -425,160 +403,11 @@ def vim_colorscheme(name, palette)
     hi 'markdownItalic'     , :markdownItalic
     hi 'markdownBoldItalic' , :markdownBoldItalic
   end
-  vim.output
 end
 
-color = {
-  red:     "\#ef7070",
-  red2:    "\#8a1a1a",
-  orange:  "\#e8a368",
-  yellow:  "\#d8cf6c",
-  yellow2: "\#e2d86a",
-  green:   "\#85c178",
-  green2:  "\#5cc46b",
-  cyan:    "\#64bac9",
-  blue:    "\#6882e2",
-  purple:  "\#b281d1",
-  gray1:   "\#222c33",
-  gray2:   "\#314047",
-  gray3:   "\#506872",
-  gray4:   "\#5f7b87",
-  gray5:   "\#7395a3",
-  white:   "\#a9afb2",
-  black:   "\#050a26",
-}
+vim = gen_vim('otynium', style_otynium)
+vim256 = gen_vim('otynium256', style_otynium256)
 
-color256 = {
-  red:     "\#d75f5f",
-  red2:    "\#870000",
-  orange:  "\#ffaf5f",
-  yellow:  "\#d7d75f",
-  yellow2: "\#d7d700",
-  green:   "\#87af87",
-  green2:  "\#5faf5f",
-  cyan:    "\#5fafd7",
-  blue:    "\#5f87ff",
-  purple:  "\#af87ff",
-  gray1:   "\#262626",
-  gray2:   "\#303030",
-  gray3:   "\#4e4e4e",
-  gray4:   "\#585858",
-  gray5:   "\#767676",
-  white:   "\#a8a8a8",
-  black:   "\#00005f",
-}
-
-def palette(color)
-  palette = Palette.new(color)
-  palette.instance_eval do
-  hi :Normal              , :white   , :gray1  , []
-  hi :NonText             , :gray2   , nil     , []
-  hi :Cursor              , :gray2   , :gray3  , [:none]
-  hi :CursorLine          , nil      , :gray2  , [:none]
-  hi :CursorLineNr        , :cyan    , :gray2  , [:none]
-  hi :CursorColumn        , nil      , :gray2  , []
-  hi :LineNr              , :gray2   , nil     , []
-  hi :StatusLine          , :gray5   , :gray2  , []
-  hi :StatusLineNC        , :gray1   , :gray4  , []
-  hi :StatusLineTerm      , :gray5   , :gray2  , []
-  hi :StatusLineTermNC    , :gray1   , :gray4  , []
-  hi :Title               , :green   , nil     , []
-  hi :Directory           , :blue    , nil     , []
-  hi :Pmenu               , :white   , :black  , []
-  hi :PmenuSel            , :white   , :gray2  , []
-  hi :PmenuSbar           , :gray3   , nil     , []
-  hi :PmenuThumb          , :gray4   , nil     , []
-  hi :Search              , :gray2   , :yellow , []
-  hi :IncSearch           , :gray1   , :yellow , []
-  hi :DiffAdd             , :cyan    , :gray2  , []
-  hi :DiffChange          , :yellow2 , :gray2  , []
-  hi :DiffDelete          , :red     , :gray2  , []
-  hi :DiffText            , :green2  , :gray2  , []
-  hi :SpellBad            , :red     , nil     , []
-  hi :SpellCap            , :blue    , nil     , []
-  hi :SpellLocal          , :green   , nil     , []
-  hi :SpellRare           , :gray2   , :yellow , []
-  hi :SpecialKey          , :gray2   , nil     , []
-  hi :WildMenu            , :blue    , nil     , []
-  hi :Folded              , :gray4   , :gray1  , []
-  hi :FoldColumn          , :gray4   , :gray1  , []
-  hi :VertSplit           , :gray4   , nil     , []
-  hi :TabLine             , :gray4   , :gray2  , []
-  hi :TabLineFill         , :gray4   , :gray2  , []
-  hi :TabLineSel          , :orange  , :gray3  , []
-  hi :ColorColumn         , nil      , :gray1  , []
-  hi :SignColumn          , :gray5   , :gray1  , []
-  hi :Question            , :cyan    , nil     , []
-  hi :Visual              , :gray5   , :gray3  , []
-  hi :VisualNOS           , :gray5   , :gray3  , []
-  hi :ModeMsg             , :yellow  , nil     , []
-  hi :MoreMsg             , :yellow  , nil     , []
-  hi :WarningMsg          , :red     , nil     , []
-  hi :ErrorMsg            , nil      , :red2   , []
-  hi :Constant            , :orange  , nil     , []
-  hi :String              , :green   , nil     , []
-  hi :Character           , :green   , nil     , []
-  hi :Number              , :orange  , nil     , []
-  hi :Float               , :orange  , nil     , []
-  hi :Boolean             , :orange  , nil     , []
-  hi :Identifier          , :red     , nil     , [:none]
-  hi :Function            , :cyan    , nil     , []
-  hi :Statement           , :purple  , nil     , []
-  hi :Conditional         , :purple  , nil     , []
-  hi :Repeat              , :purple  , nil     , []
-  hi :Label               , :purple  , nil     , []
-  hi :Operator            , :blue    , nil     , []
-  hi :Keyword             , :purple  , nil     , []
-  hi :Exception           , :yellow  , nil     , []
-  hi :Type                , :yellow  , nil     , []
-  hi :StorageClass        , :orange  , nil     , []
-  hi :Structure           , :orange  , nil     , []
-  hi :Typedef             , :yellow  , nil     , []
-  hi :PreProc             , :cyan    , nil     , []
-  hi :Include             , :cyan    , nil     , []
-  hi :Define              , :cyan    , nil     , []
-  hi :Macro               , :cyan    , nil     , []
-  hi :PreCondit           , :blue    , nil     , []
-  hi :Special             , :cyan    , nil     , []
-  hi :SpecialChar         , :cyan    , nil     , []
-  hi :Tag                 , :blue    , nil     , []
-  hi :Delimiter           , :blue    , nil     , []
-  hi :SpecialComment      , :cyan    , nil     , []
-  hi :Debug               , :red     , nil     , []
-  hi :Underlined          , :red     , nil     , []
-  hi :Comment             , :gray4   , nil     , []
-  hi :Todo                , :gray1   , :yellow , []
-  hi :Ignore              , :gray3   , nil     , []
-  hi :Error               , nil      , :red2   , []
-  hi :MatchParen          , :gray2   , :cyan   , []
-  hi :Warning             , :orange  , nil     , []
-  hi :hint                , nil      , :gray3  , []
-  hi :hintMsg             , :gray3   , nil     , []
-  hi :Info                , nil      , :cyan   , []
-  hi :InfoMsg             , :cyan    , nil     , []
-  hi :CocErrorhighlight   , nil      , nil     , [:underline]
-  hi :CocWarninghighlight , nil      , nil     , [:underline]
-  hi :Cochinthighlight    , nil      , nil     , [:underline]
-  hi :CocInfohighlight    , nil      , nil     , [:underline]
-  hi :TSVariable          , :cyan    , nil     , []
-  hi :TSStrong            , :white   , :gray1  , [:bold]
-  hi :TSStrong            , :white   , :gray1  , [:underline]
-  hi :TSLiteral           , :orange  , nil     , []
-  hi :TSURI               , :orange  , nil     , []
-  hi :TSTagDelimiter      , :gray4   , nil     , []
-  hi :markdownH1          , :white   , nil     , [:bold]
-  hi :markdownH2          , :white   , nil     , [:bold]
-  hi :markdownH3          , :white   , nil     , [:bold]
-  hi :markdownH4          , :white   , nil     , [:bold]
-  hi :markdownH5          , :white   , nil     , [:bold]
-  hi :markdownH6          , :white   , nil     , [:bold]
-  hi :markdownBold        , :white   , nil     , [:bold]
-  hi :markdownItalic      , :white   , nil     , [:italic]
-  hi :markdownBoldItalic  , :white   , nil     , [:italic     , :bold]
-  end
-  palette
-end
-
-File.write("vim/otynium.vim", vim_colorscheme('otynium', palette(color)))
-File.write("vim/otynium256.vim", vim_colorscheme('otynium256', palette(color256)))
-File.write("vscode/themes/otynium-color-theme.json", vscode_colorscheme('otynium', palette(color), 'dark'))
+File.write("vim/otynium.vim", vim)
+File.write("vim/otynium256.vim", vim256)
+File.write("vscode/themes/otynium-color-theme.json", vscode)
